refactor: restructure PostCard component and move to separate file
